+++ couplet_composer/support/file_paths.py
@@ -32,6 +32,14 @@ def get_project_values_file_path():
     return os.path.join("util", "project.json")
 
 
+def get_product_file_path():
+    """
+    Gives the path to the file that contains constants of the
+    project this script acts on relative to the Ode repository.
+    """
+    return "product.json"
+
+
 def get_project_dependencies_file_path(source_root):
     """
     Gives the path to the file that contains constants of the
@@ -43,14 +51,24 @@ def get_project_dependencies_file_path(source_root):
     """
     old_file = os.path.join("util", "dependencies.json")
     if os.path.isfile(os.path.join(source_root, old_file)):
-        logging.warn(
+        logging.warning(
             "The file '%s' for providing dependency information is "
             "deprecated; use %s instead",
             old_file,
             get_project_values_file_path()
         )
         return old_file
-    return get_project_values_file_path()
+    if os.path.isfile(
+        os.path.join(source_root, get_project_values_file_path())
+    ):
+        logging.warning(
+            "The file '%s' for providing dependency information is "
+            "deprecated; use %s instead",
+            get_project_values_file_path(),
+            get_product_file_path()
+        )
+        return get_project_values_file_path()
+    return get_product_file_path()
 
 
 def get_github_api_file_path():